+++ src/Audit.NET.PostgreSql/ConfigurationApi/IPostgreSqlProviderConfigurator.cs
@@ -45,7 +45,7 @@ namespace Audit.PostgreSql.Configuration
         /// <param name="dataColumnName">The data column name.</param>
         /// <param name="dataColumnType">The data column type.</param>
         /// <param name="jsonStringBuilder">A function that returns the JSON string to store in the data column. By default it's the result of calling AuditEvent.ToJson().</param>
-        IPostgreSqlProviderConfigurator DataColumn(string dataColumnName, DataType dataColumnType = DataType.JSON,Func<AuditEvent, string> jsonStringBuilder = null);
+        IPostgreSqlProviderConfigurator DataJsonColumn(string dataColumnName, DataType dataColumnType = DataType.JSON,Func<AuditEvent, string> jsonStringBuilder = null);
 
         /// <summary>
         /// Specifies the column name and type where to store the event data. Default is the column named "data" assuming type JSON.
@@ -53,7 +53,7 @@ namespace Audit.PostgreSql.Configuration
         /// <param name="dataColumnNameBuilder">The data column name builder.</param>
         /// <param name="dataColumnType">The data column type.</param>
         /// <param name="jsonStringBuilder">A function that returns the JSON string to store in the data column. By default it's the result of calling AuditEvent.ToJson().</param>
-        IPostgreSqlProviderConfigurator DataColumn(Func<AuditEvent, string> dataColumnNameBuilder, DataType dataColumnType = DataType.JSON, Func<AuditEvent, string> jsonStringBuilder = null);
+        IPostgreSqlProviderConfigurator DataJsonColumn(Func<AuditEvent, string> dataColumnNameBuilder, DataType dataColumnType = DataType.JSON, Func<AuditEvent, string> jsonStringBuilder = null);
 
         /// <summary>
         /// Specifies the column where to store the last updated date. NULL to ignore. Default is NULL.